Catalog: move bucket name/stats/create-package above the tabs (visible on all tabs); Overview keeps description/readme/charts
